+++ changelog.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v8.1.0...HEAD)
+* [PCM-2108](https://inindca.atlassian.net/browse/PCM-2108) - fix `sdk.acceptSession` so it doesn't send the request to the backend more than once.
+* [PCM-2162](https://inindca.atlassian.net/browse/PCM-2162) - fix `sessionStarted` events so they only get emitted once. (only affected softphone)
+
 # [v8.1.0](https://github.com/MyPureCloud/genesys-cloud-webrtc-sdk/compare/v8.0.11...v8.1.0)
 ### Added
 * [PCM-2114](https://inindca.atlassian.net/browse/PCM-2114) - Allow sessions to be reestablished through a reinvite process.